requirements.txt
typer>=0.9.0
pandas>=2.0
numpy>=1.26
biopython>=1.80
gffutils>=0.10.1
matplotlib>=3.7
openpyxl>=3.1
requests>=2.31